Sketch RC1

+++ VisualBasic/JoloFocuser/JoloFocuser/My Project/Settings.Designer.vb
@@ -105,9 +105,9 @@ Namespace My
         <Global.System.Configuration.UserScopedSettingAttribute(),  _
          Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
          Global.System.Configuration.DefaultSettingValueAttribute("0")>  _
-        Public Property StepsPerC() As Decimal
+        Public Property StepsPerC() As Double
             Get
-                Return CType(Me("StepsPerC"),Decimal)
+                Return CType(Me("StepsPerC"),Double)
             End Get
             Set
                 Me("StepsPerC") = value
